Fix by review
- Move the ReadKind to the AbsolutePath from the LocalPath
- Remove the WriteAttributes from the CreateFile
